updated everything to have Feed parent component which passes values as props to the streak counter and input components
