+++ frontend/components/input/Input.jsx
@@ -1,15 +1,12 @@
-import { View, Text, Button} from 'react-native'
 import React, { useState} from 'react';
-import Total from "../../classes/Total";
-import StreakCounter from '../streak-counter/StreakCounter';
-import { setTotalExternal } from '../streak-counter/StreakCounter';
+import { View, Button} from 'react-native'
 
-const Input = () => {
+const Input = ({ handleMealDealConsumed }) => {
  
-
   const HandlePress=()=>{
-     setTotalExternal() 
+     handleMealDealConsumed()
   }
+
   return (
     <View>
       <Button title='Add Meal Dealio' onPress={HandlePress}/>
